First stab at moving stuff to `data` package.
