+++ packages/data/src/index.ts
@@ -1,7 +1,25 @@
 /**
  * @license
- * Copyright 2025 Google LLC
+ * Copyright 2024 Google LLC
  * SPDX-License-Identifier: Apache-2.0
  */
+import { DataStore, RunStore } from "@breadboard-ai/types";
+import { DefaultDataStore } from "./default-data-store.js";
+import { DefaultRunStore } from "./default-run-store.js";
 
-console.log("code goes here");
+export const createDefaultDataStore = (): DataStore => {
+  return new DefaultDataStore();
+};
+
+export const createDefaultRunStore = (): RunStore => {
+  return new DefaultRunStore();
+};
+
+export {
+  deflateData,
+  inflateData,
+  purgeStoredDataInMemoryValues,
+  transformContents,
+  visitGraphNodes,
+  remapData,
+} from "./inflate-deflate.js";
