+++ packages/runtime/package.json
@@ -129,6 +129,7 @@
     "typescript": "^5.8.3"
   },
   "dependencies": {
+    "@breadboard-ai/data": "0.0.1",
     "@breadboard-ai/loader": "0.0.1",
     "@breadboard-ai/types": "0.7.0",
     "@breadboard-ai/utils": "0.0.1"
