+++ packages/runtime/src/harness/local.ts
@@ -15,12 +15,12 @@ import {
   RunConfig,
 } from "@breadboard-ai/types";
 import { asyncGen } from "@breadboard-ai/utils";
-import { createDefaultDataStore } from "../data/index.js";
 import { runGraph } from "../run/run-graph.js";
 import { timestamp } from "../timestamp.js";
 import { Diagnostics } from "./diagnostics.js";
 import { extractError } from "./error.js";
 import { baseURL } from "./url.js";
+import { createDefaultDataStore } from "@breadboard-ai/data";
 
 const fromProbe = <Probe extends ProbeMessage>(probe: Probe) => {
   const data = structuredClone(probe.data);
